Moved errors to pkg and added comments.
