+++ cmd/basicgopot/main.go
@@ -4,8 +4,8 @@ import (
 	"os"
 
 	"github.com/morgenm/basicgopot/internal/config"
-	"github.com/morgenm/basicgopot/internal/errors"
 	"github.com/morgenm/basicgopot/internal/server"
+	"github.com/morgenm/basicgopot/pkg/errors"
 )
 
 func main() {
