+++ internal/server/checkVt.go
@@ -8,7 +8,7 @@ import (
 	"time"
 
 	"github.com/morgenm/basicgopot/internal/config"
-	"github.com/morgenm/basicgopot/internal/errors"
+	"github.com/morgenm/basicgopot/pkg/errors"
 	"github.com/morgenm/basicgopot/pkg/vt"
 )
